inherit SampleController from OpenController
