+++ lib/wab/open_controller.rb
@@ -1,14 +1,24 @@
 
 module WAB
 
-  # A Controller that makes all CRUD methods public. The handle method raises
-  # as it has to be handled specially.
+  # This controller exposes all possible methods expected in a WAB::Controller
+  # subclass, as public methods.
+  #
+  # Since those methods are private in the superclass, they need to be redefined
+  # as public methods to enable the concerned functionality.
+  #
+  # For example, if a controller is intented to provide only read-access, then
+  # just the +read+ method would need to be exposed as a public method. The
+  # remaining methods may remain private.
   class OpenController < Controller
 
     def initialize(shell)
       super(shell)
     end
 
+    # Catch requests that are not one of the below CRUD methods.
+    #
+    # Raises as it has to be handled specially.
     def handle(data)
       raise NotImplementedError.new
     end
